Deployed version for everyone
Finally, a deployed version for everyone! This creates a web server with a simple interface for using the final model. The steps you need to follow are quite simple: uncompress the folder, execute app.py, and go to the link!

The only packages that need to be installed are NumPy and Flask.